+++ docs/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ### ✨ Features
 
-* Crawl Tiktok Affiliate Creator ([c8b9d2a](https://github.com/hdevlinz/tiktok-affiliate-crawler/commit/c8b9d2a151b1f36c0dc0d1ac261637cfe75cf316))
+- Crawl Tiktok Affiliate Creator ([c8b9d2a](https://github.com/hdevlinz/tiktok-affiliate-crawler/commit/c8b9d2a151b1f36c0dc0d1ac261637cfe75cf316))